+++ lib/elixir_script/passes/filter_exjs.ex
@@ -1,0 +1,14 @@
+defmodule ElixirScript.Passes.FilterExjs do
+
+
+  def execute(compiler_data, _) do
+
+    data = Enum.filter(compiler_data.data, fn
+      {_, data} -> Path.extname(data.path) == ".exjs"
+      data -> Path.extname(data.path) == ".exjs"
+    end)
+
+    Map.put(compiler_data, :data, data)
+  end
+
+end