+++ lib/elixir_script/passes/load_modules.ex
@@ -1,0 +1,26 @@
+defmodule ElixirScript.Passes.LoadModules do
+  @pass 5
+
+  def execute(compiler_data, _) do
+    ex_files = Enum.map(compiler_data.data, fn
+      { _, %{path: path} } -> path
+      %{path: path} -> path
+    end)
+    |> Enum.filter(fn path -> Path.extname(path) == ".ex" || Path.extname(path) == ".exs" end)
+
+    loaded_modules = case ex_files do
+      [] ->
+        []
+      files ->
+        try do
+          Kernel.ParallelRequire.files(files)
+        rescue
+          _ ->
+            []
+        end
+    end
+
+    Map.put(compiler_data, :loaded_modules, loaded_modules)
+  end
+
+end