+++ lib/elixir_script/translator.ex
@@ -28,7 +28,6 @@ defmodule ElixirScript.Translator do
   alias ESTree.Tools.Builder, as: JS
   alias ElixirScript.Translator.Rewriter
   alias ElixirScript.Translator.Receive
-  alias ElixirScript.Translator.LexicalScope
 
   # A list of erlang modules. These are rewritten into equivalent
   # JavaScript functions using ElixirScript.Translator.Rewriter
@@ -539,10 +538,7 @@ defmodule ElixirScript.Translator do
   end
 
   defp do_translate({name, _, params} = ast, env) when is_list(params) do
-
-
       expanded_ast = Macro.expand(ast, env.env)
-
       if expanded_ast == ast do
         name_arity = {name, length(params)}
         module = ElixirScript.Translator.State.get_module(env.module)