+++ test/compiler/deps_test.exs
@@ -1,5 +1,0 @@
-defmodule ElixirScript.Compiler.Deps.Test do
-  use ExUnit.Case
-  alias  ElixirScript.Compiler.Cache
-
-end